requirements.txt
python-dotenv
requests
discord.py
discord.ext
os
logging
pexels_api



// Paramètres originaux : 
// {
//     "editor.fontLigatures": true,
//     "editor.cursorBlinking": "smooth",
//     "editor.cursorSmoothCaretAnimation": "on",
//     "liveServer.settings.donotVerifyTags": true,
//     "python.defaultInterpreterPath": "C:\\Users\\tbenn\\AppData\\Local\\Microsoft\\WindowsApps\\python3.12.exe",
//     "git.autofetch": true,
//     "files.autoSave": "afterDelay",
//     "workbench.colorTheme": "Bearded Theme Black & Ruby",
//     "workbench.iconTheme": "bearded-icons",
//     "security.workspace.trust.untrustedFiles": "open", 
//     "window.opacity": 0.85,
// }


// Parmaètres actuels :
{
    "editor.fontLigatures": true,
    "editor.fontFamily": "JetBrains Mono, Fira Code, Cascadia Code, Consolas",
    "editor.fontSize": 14,
    "editor.lineHeight": 1.6,
    "editor.cursorBlinking": "smooth",
    "editor.cursorSmoothCaretAnimation": "on",
    "editor.smoothScrolling": true,
    "editor.minimap.enabled": true,
    "editor.bracketPairColorization.enabled": true,
    "editor.guides.bracketPairs": true,
    
    "liveServer.settings.donotVerifyTags": true,
    "python.defaultInterpreterPath": "C:\\Users\\tbenn\\AppData\\Local\\Microsoft\\WindowsApps\\python3.12.exe",
    
    "git.autofetch": true,
    "git.confirmSync": false,
    "git.enableSmartCommit": true,
    
    "files.autoSave": "afterDelay",
    "files.autoSaveDelay": 1000,
    
    "workbench.colorTheme": "Bearded Theme Black & Ruby",
    "workbench.iconTheme": "bearded-icons",
    "workbench.startupEditor": "none",
    "workbench.tree.indent": 20,
    
    "security.workspace.trust.untrustedFiles": "open", 
    
    "window.opacity": 0.85,
    "window.title": "${dirty}${activeEditorShort}${separator}${rootName}",
    "window.titleBarStyle": "custom",
    
    "terminal.integrated.fontSize": 13,
    "terminal.integrated.fontFamily": "JetBrains Mono, Cascadia Code",
    "terminal.integrated.cursorBlinking": true,
    "terminal.integrated.cursorStyle": "line",

    "workbench.colorCustomizations": {
        },
    "window.controlsStyle": "custom",
}